+++ driver/key/keyM.c
@@ -17,7 +17,7 @@ static csr_keym_cfg_t csr_keym_cfg = {
 	.last_state = KEY_M_UP,
 };
 
-s16 csr_keym_event_handler(u16 key_num, u16 key_status)
+s16 csr_keym_event_handler(u32 key_num, u32 key_status)
 {
 	EVENT_E now_state;
 	if(key_num & (1UL << csr_keym_cfg.pin.num)) {
